move common functions to pkg/metrics.
