+++ pd-client/rpc_worker.go
@@ -29,6 +29,7 @@ import (
 	"github.com/pingcap/kvproto/pkg/msgpb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	"github.com/pingcap/kvproto/pkg/util"
+	"github.com/pingcap/pd/pkg/metrics"
 	"github.com/twinj/uuid"
 )
 
@@ -315,7 +316,7 @@ func (w *rpcWorker) getClusterConfigFromRemote(conn *bufio.ReadWriter, clusterCo
 func (w *rpcWorker) callRPC(conn *bufio.ReadWriter, req *pdpb.Request) (resp *pdpb.Response, err error) {
 	// Record some metrics.
 	start := time.Now()
-	label := convertName(req.GetCmdType().String())
+	label := metrics.GetCmdLabel(req)
 	defer func() {
 		if err == nil {
 			cmdCounter.WithLabelValues(label).Inc()
